requirements.txt
-i https://pypi.org/simple
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
deprecated==1.2.13
django==4.1.2
django-adminlte3==0.1.6
django-crispy-forms==1.14.0
django-cron==0.6.0
django-filter==22.1
django-jazzmin==2.5.0
djangorestframework==3.14.0
install==1.3.5
kombu==5.2.4
mysql-connector-python==8.0.30
packaging==21.3
pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.20.1
pymysql==1.0.2
pyparsing==3.0.9
pytz==2022.4
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
